requirements.txt
Flask==1.1.1
google-cloud-vision==0.41.0
google-cloud-storage
google-cloud-ndb